Move formatTime in utils. Move graph URL to top scope. Short circuit account & chainId check. Use destructuring assignment. Add a variable array to store unsorted txn.
